Check library availability before loading
git-svn-id: http://jira.uji.es/repos/uji/CRYPTOAPPLET/branches/CryptoApplet_2.1.4-detached@16405 e6edf6fb-f266-4316-afb4-e53d95876a76

+++ uji-keystore/src/main/java/es/uji/security/keystore/mozilla/Mozilla.java
@@ -153,7 +153,10 @@ public class Mozilla
     			{
     				for (String orig:libraries)
     				{
-    					OS.copyfile(res + "\\" + orig, strTmpDir +  "\\" + orig);
+    				    if (new File(res + "\\" + orig).exists())
+    				    {
+    				        OS.copyfile(res + "\\" + orig, strTmpDir +  "\\" + orig);
+    				    }
     				}
     				res= strTmpDir; 
     			}
